+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Renamed `modeButtons` configuration field to `naturalModeButton`
+- the new `naturalModeButton` configuration field now creates just one switch which enables/disabled the natural mode
+- fixed the status of the rotation direction button(natural mode)
+
 
 ## [1.0.2] - 2020-07-08
 ### Changed